fix the CI image builds again + some flutter deployment related changes
